# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Affordability Calculator",
    "description": "Mortgage calculation tools for home buyers. Calculate affordability, analyze properties, compare loan scenarios, stress-test finances, and get live market data.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://aicalculator.homes"
    }
  ],
  "paths": {
...
    "/api/gpt/search-properties": {
      "post": {
        "operationId": "searchProperties",
        "summary": "Search for homes currently for sale in a US city",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["location"],
                "properties": {
                  "location": {
                    "type": "string",
                    "description": "City and state (e.g., 'Austin, TX')"
                  },
                  "max_price": {
                    "type": "number",
                    "description": "Maximum listing price filter"
                  },
                  "min_beds": {
                    "type": "integer",
                    "description": "Minimum number of bedrooms"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property listings from Zillow",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": { "type": "string" },
                    "resultCount": { "type": "integer" },
                    "listings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "address": { "type": "string" },
                          "price": { "type": "number" },
                          "beds": { "type": "integer" },
                          "baths": { "type": "number" },
                          "sqft": { "type": "number" },
                          "homeType": { "type": "string" },
                          "listingUrl": { "type": "string" }
                        }
                      }
                    },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}